+++ lib/i18n/strings.g.dart
@@ -4,9 +4,9 @@
 /// To regenerate, run: `dart run slang`
 ///
 /// Locales: 1
-/// Strings: 800
+/// Strings: 825
 ///
-/// Built on 2026-06-11 at 18:27 UTC
+/// Built on 2026-06-11 at 21:15 UTC
 
 // coverage:ignore-file
 // ignore_for_file: type=lint, unused_import